move integration test to test_provider.py
